requirements.txt
pandas
apache-airflow-providers-amazon
airflow-ai-sdk[groq]